Create products.jsx
This Products page displays a grid of product cards using the Material-UI grid system. Each product card contains an image, a name, and a description of the product. The product data is stored in an array of objects, but it can also be retrieved from a database or an API.